Modularization - 1 - Seperated Login form from authentication
