+++ src/pages/Login/LoginForm.jsx
@@ -1,0 +1,70 @@
+import React, { useState } from 'react';
+import { Mail, Lock } from 'lucide-react';
+
+function LoginForm({ handleLogin, loading, navigate }) {
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+
+  const onSubmit = (e) => {
+    e.preventDefault();
+    handleLogin(email, password);
+  };
+
+  return (
+    <form onSubmit={onSubmit} className="space-y-6">
+      <div>
+        <label htmlFor="email" className="block text-sm font-medium text-gray-700">
+          Email Address
+        </label>
+        <div className="mt-1 relative">
+          <Mail className="absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-400 h-5 w-5" />
+          <input
+            id="email"
+            type="email"
+            required
+            className="pl-10 block w-full rounded-md border-gray-300 shadow-sm focus:border-[#101d42] focus:ring focus:ring-[#101d42] focus:ring-opacity-50"
+            value={email}
+            onChange={(e) => setEmail(e.target.value)}
+          />
+        </div>
+      </div>
+
+      <div>
+        <label htmlFor="password" className="block text-sm font-medium text-gray-700">
+          Password
+        </label>
+        <div className="mt-1 relative">
+          <Lock className="absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-400 h-5 w-5" />
+          <input
+            id="password"
+            type="password"
+            required
+            className="pl-10 block w-full rounded-md border-gray-300 shadow-sm focus:border-[#101d42] focus:ring focus:ring-[#101d42] focus:ring-opacity-50"
+            value={password}
+            onChange={(e) => setPassword(e.target.value)}
+          />
+        </div>
+      </div>
+
+      <button
+        type="submit"
+        disabled={loading}
+        className="w-full flex justify-center py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-[#101d42] hover:bg-opacity-90 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-[#101d42] disabled:opacity-50"
+      >
+        {loading ? 'Logging in...' : 'Login'}
+      </button>
+
+      <div className="text-center">
+        <button
+          type="button"
+          onClick={() => navigate('/signup')}
+          className="text-sm text-[#101d42] hover:text-opacity-90"
+        >
+          Don't have an account? Sign up
+        </button>
+      </div>
+    </form>
+  );
+}
+
+export default LoginForm;
